Pulled out RigidBodyActuator into its own files.
