+++ drake/systems/plants/CMakeLists.txt
@@ -20,6 +20,7 @@ add_library_with_exports(LIB_NAME drakeRBM SOURCE_FILES
   RigidBodyTree.cpp
   RigidBody.cpp
   RigidBodyTreeContact.cpp
+  rigid_body_actuator.cc
   rigid_body_frame.cc)
 target_link_libraries(drakeRBM drakeCollision drakeJoints drakeUtil drakeXMLUtil)
 pods_install_libraries(drakeRBM)
@@ -30,6 +31,7 @@ drake_install_headers(
   KinematicsCache.h
   material_map.h
   pose_map.h
+  rigid_body_actuator.h
   RigidBodyFrame.h
   RigidBody.h
   RigidBodyTree.h
